modified bash scripts for backend

+++ src/backend/HMT/start.sh
@@ -1,9 +1,28 @@
+# 
+
 echo "Activating virtual environment"
 . venv/bin/activate
 
-echo "Installing dependencies"
-pip install --upgrade pip
-pip install -r requirements.txt --no-cache-dir
+# original script to install dependencies
+# echo "Installing dependencies"
+# pip install --upgrade pip
+# pip install -r requirements.txt --no-cache-dir
+
+
+# added choice to install dependencies because it takes
+# a while to download - after the first time of installing,
+# skip this step
+read -p "Dependencies already installed? (y|n): " res
+
+if [ $res = "n" ]; then
+    pip install --upgrade pip
+    pip install -r requirements.txt --no-cache-dir
+fi
+
+# Could be for windows, double check at meeting
+# echo "Starting service"
+# python3 -m src.app &
 
 echo "Starting service"
-python -m src.app &
+cd src
+python3 app.py &
